Release 1.24 - 24/04/2025
Updated HE Code snapshot with latest prod version

+++ src/Web App/HE.Remediation.Core/UseCase/Areas/FireRiskAppraisal/CheckYourAnswers/GetCheckYourAnswers/GetCheckYourAnswersResponse.cs
@@ -23,7 +23,7 @@ public class GetCheckYourAnswersResponse
 
     public int? NumberOfStoreys { get; set; }
 
-    public int? BuildingHeight { get; set; }
+    public decimal? BuildingHeight { get; set; }
 
     public EBasicComplexType? BasicComplexType { get; set; }
