+++ src/Web App/HE.Remediation.WebApp/ViewModels/FireRiskAppraisal/CheckYourAnswersViewModel.cs
@@ -25,7 +25,7 @@ namespace HE.Remediation.WebApp.ViewModels.FireRiskAppraisal
 
         public int? NumberOfStoreys { get; set; }
 
-        public int? BuildingHeight { get; set; }
+        public decimal? BuildingHeight { get; set; }
 
         public EBasicComplexType? BasicComplexType { get; set; }
